% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for.R
\name{search_for}
\alias{search_for}
\title{Search for similar sentences}
\usage{
search_for(x, sen, exact = FALSE, cos_sim = 0.5)
}
\arguments{
\item{x}{File name/path of the PDF.}

\item{sen}{Sentence to be used to search in the text.}

\item{exact}{If you search for the exact sentence, the default is FALSE and the
cosine distance is used as similarity measurement.}

\item{cos_sim}{Similarity parameter of the cosine distance. The output contains
sentences which have cosine similarity greater or equal 'cos_sim'. The default
is 0.5.}
}
\value{
A tibble data frame that contains the measured cosine similarity and
the location of the match, the page number and the sentence number.
}
\description{
\code{search_for} is used to search for similar or exact sentences in a PDF.
}
\examples{
# PDF from Book Reports,
# URL: https://www.bookreports.info/hansel-and-gretel-summary/
file <- system.file('pdf', 'summary_hansel_and_gretel.pdf', package = 'antiplugr')

# a similar sentence from 'grimm_hanse_and_gretel.pdf' from Short Story America,
# URL: http://www.shortstoryamerica.com/pdf_classics/grimm_hanse_and_gretel.pdf
sen_1 <- "When four weeks had passed and Hansel was still thin, impatience
overcame her, and she would wait no longer."

# an exact sentences
sen_2 <- "When four weeks had passed and Hansel was still thin, the witch
got tired."

search_for(file, sen_1)
search_for(file, sen_2, exact = TRUE)

}
